add video owner, (edit, delete) authority

+++ src/controllers/videoController.js
@@ -1,8 +1,5 @@
 import Video from "../models/Video";
-
-// Video.find({}, (error, videos) => {
-//     res.render("home", { pageTitle: "Home", videos });
-// });
+import User from "../models/User";
 
 export const home = async (req, res, next) => {
     try {
@@ -12,23 +9,35 @@ export const home = async (req, res, next) => {
         return res.status(400).render("server-error");
     }
 }
+
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
+
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
     return res.render("videos/watch", { pageTitle: video.title, video });
 }
-    
+
+// edit video
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id }
+    } = req.session;
+
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." })
     }
+    
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     return res.render("videos/edit", { pageTitle: `Edit: ${video.title}`, video });
 }
+
 export const postEdit = async (req, res) => {
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
@@ -44,20 +53,29 @@ export const postEdit = async (req, res) => {
     return res.redirect(`/videos/${id}`);
 }
 
+// upload video
 export const getUpload = (req, res) => {
     return res.render("videos/upload", { pageTitle: "Upload Video" });
 }
 
 export const postUpload = async (req, res) => {
-    const { path: fileUrl } = req.file;
-    const { title, description, hashtags } = req.body;
+    const { 
+        session: { user: { _id } },
+        file: { path: fileUrl },
+        body: { title, description, hashtags }
+    } = req;
+
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags)
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
     } catch(error) {
         return res.status(400).render("videos/upload", { pageTitle: "Upload Video", errorMessage: error._message });
     }
@@ -66,6 +84,18 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id }
+    } = req.session;
+
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found." })
+    }
+
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
